Export env variables into separate file
- Fix SubsocialApi IPFS connection
