+++ package.json
@@ -92,9 +92,9 @@
     "@polkadot/typegen": "1.11.1",
     "@polkadot/types": "1.11.1",
     "@polkadot/ui-assets": "^0.52.0-beta.3",
-    "@subsocial/api": "^0.2.16",
-    "@subsocial/types": "^0.1.29",
-    "@subsocial/utils": "^0.1.21",
+    "@subsocial/api": "^0.2.19",
+    "@subsocial/types": "^0.1.30",
+    "@subsocial/utils": "^0.1.23",
     "@zeit/next-css": "^1.0.1",
     "antd": "^3.25.0",
     "antd-mobile": "^2.3.1",
